refactor: group the drawer player under components/player
The player was spread across src and src/components with names that
carried the grouping instead of the directory: playerLoader,
nativePlayerBuilder, playerDrawer, playerIcons, drawerDom,
trackSelection.

They now sit together in components/player as loader, builder, drawer,
icons, query and trackSelection, with an index exposing what the rest
of the codebase actually uses. label_view and continuousPlay import
from components/player rather than reaching for individual files.

discography stays outside: it reads the page rather than playing
anything. keyboardShortcuts stays outside too, since the album page
player shares it. The top level player.ts also stays, being page level
wiring rather than part of the drawer.

drawerDom becomes query rather than dom, since components already has a
dom module and '../dom' beside './dom' would invite misreading.
