+++ src/components/player/icons.ts
