+++ src/components/player/index.ts
@@ -1,0 +1,2 @@
+export { createPlayerDrawer } from './drawer';
+export { loadAlbumIntoDrawer, loadNextAlbum, loadPreviousAlbum } from './loader';
