+++ src/components/player/query.ts
